user management final part

+++ ElectroGridSystem/src/main/java/electro/controller/userController.java
@@ -6,8 +6,8 @@ import javax.ws.rs.core.MediaType;
 import com.google.gson.*;
 import org.json.simple.*;
 
+import electro.classes.*;
 import electro.model.*;
-import electro.service.*;
 
 import org.apache.tomcat.util.json.JSONParser;
 import org.jsoup.*;
@@ -34,23 +34,33 @@ public class userController {
 		
 		//Convert the input string to a JSON object
 		JsonObject app = new JsonParser().parse(app_text).getAsJsonObject();
+
+		if(app.get("name").getAsString()!=""&&app.get("email").getAsString()!=""&&app.get("nic").getAsString()!=""&&app.get("phone").getAsString()!=""&&app.get("address").getAsString()!=""&&app.get("pass").getAsString()!=""&&app.get("privilege").getAsString()!="") {
 		
-		User user = new User();
-		user.setName(app.get("name").getAsString());
-		user.setEmail(app.get("email").getAsString());
-		user.setNic(app.get("nic").getAsString());
-		user.setPhone(app.get("phone").getAsString());
-		user.setAddress(app.get("address").getAsString());
-		user.setPassword(app.get("pass").getAsString());
-		user.setPrivilege(app.get("privilege").getAsString());
-		
-		user_Service.addUser(user);
-		
-		JSONObject json = new JSONObject();
-		json.put("success", user_Service.getSuccess());
-		
-		return json.toString();
+			User user = new User();
+			user.setName(app.get("name").getAsString());
+			user.setEmail(app.get("email").getAsString());
+			user.setNic(app.get("nic").getAsString());
+			user.setPhone(app.get("phone").getAsString());
+			user.setAddress(app.get("address").getAsString());
+			user.setPassword(app.get("pass").getAsString());
+			user.setPrivilege(app.get("privilege").getAsString());
 			
+			user_Service.addUser(user);
+			
+			JSONObject json = new JSONObject();
+			json.put("success", user_Service.getSuccess());
+			
+			return json.toString();
+
+		}else {
+			
+			JSONObject json = new JSONObject();
+			json.put("success", "Required Error");
+			
+			return json.toString();
+			
+		}			
 		
 	}
 	
@@ -63,23 +73,33 @@ public class userController {
 		//Convert the input string to a JSON object
 		JsonObject app = new JsonParser().parse(app_text).getAsJsonObject();
 
-		User user = new User();
-		user.setId(Integer.parseInt(app.get("id").getAsString()));
-		user.setName(app.get("name").getAsString());
-		user.setEmail(app.get("email").getAsString());
-		user.setNic(app.get("nic").getAsString());
-		user.setPhone(app.get("phone").getAsString());
-		user.setAddress(app.get("address").getAsString());
-		user.setPassword(app.get("pass").getAsString());
-		user.setPrivilege(app.get("privilege").getAsString());
+		if(app.get("id").getAsString()!=""&&app.get("name").getAsString()!=""&&app.get("email").getAsString()!=""&&app.get("nic").getAsString()!=""&&app.get("phone").getAsString()!=""&&app.get("address").getAsString()!=""&&app.get("pass").getAsString()!=""&&app.get("privilege").getAsString()!="") {
 		
-		user_Service.editUser(user);
-		
-		JSONObject json = new JSONObject();
-		json.put("success", user_Service.getSuccess());
-		
-		return json.toString();
+			User user = new User();
+			user.setId(Integer.parseInt(app.get("id").getAsString()));
+			user.setName(app.get("name").getAsString());
+			user.setEmail(app.get("email").getAsString());
+			user.setNic(app.get("nic").getAsString());
+			user.setPhone(app.get("phone").getAsString());
+			user.setAddress(app.get("address").getAsString());
+			user.setPassword(app.get("pass").getAsString());
+			user.setPrivilege(app.get("privilege").getAsString());
 			
+			user_Service.editUser(user);
+			
+			JSONObject json = new JSONObject();
+			json.put("success", user_Service.getSuccess());
+			
+			return json.toString();
+
+		}else {
+			
+			JSONObject json = new JSONObject();
+			json.put("success", "Required Error");
+			
+			return json.toString();
+			
+		}		
 	}
 	
 	@DELETE
@@ -90,14 +110,24 @@ public class userController {
 		
 		//Convert the input string to a JSON object
 		JsonObject app = new JsonParser().parse(app_text).getAsJsonObject();
-		
-		user_Service.deleteUser(Integer.parseInt(app.get("id").getAsString()));
-		
-		JSONObject json = new JSONObject();
-		json.put("success", user_Service.getSuccess());
-		
-		return json.toString();
+
+		if(app.get("id").getAsString()!="") {
+				
+			user_Service.deleteUser(Integer.parseInt(app.get("id").getAsString()));
 			
+			JSONObject json = new JSONObject();
+			json.put("success", user_Service.getSuccess());
+			
+			return json.toString();
+
+		}else {
+			
+			JSONObject json = new JSONObject();
+			json.put("success", "Required Error");
+			
+			return json.toString();
+			
+		}		
 	}
 	
 }
